Add files via upload
This is a backoff protocol I completed for the final project of my Data Communications class.  The specific elements required for my professor have been omitted but this code and manual shows significantly what the results found were.  This was programmed in C++ and contains the additions for honor and grad students.

I received 98% credit